pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "claude-daemon"
version = "0.1.0"
description = "Persistent daemon wrapper for Claude Code with memory, scheduling, and messaging integrations"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
dependencies = [
    "pyyaml>=6.0",
    "apscheduler>=3.10,<4",
    "python-dotenv>=1.0",
    "httpx>=0.27",
]

[project.optional-dependencies]
telegram = ["python-telegram-bot>=21.0"]
discord = ["discord.py>=2.3"]
all = [
    "python-telegram-bot>=21.0",
    "discord.py>=2.3",
]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "ruff>=0.4",
]

[project.scripts]
claude-daemon = "claude_daemon.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/claude_daemon"]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
